refactor(#59) : DB Mapper -> Entity Factory Method
