+++ src/main/java/com/blog/tech/domain/comment/entity/Comment.java
@@ -1,7 +1,8 @@
 package com.blog.tech.domain.comment.entity;
 
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 import com.blog.tech.domain.comment.dto.request.CommentRequest;
@@ -43,6 +44,20 @@ public class Comment extends BaseEntity {
 			.status(Status.REGISTERED)
 			.createdAt(LocalDateTime.now())
 			.updatedAt(LocalDateTime.now())
+			.build();
+	}
+
+	public static Comment from(final ResultSet rs, final int i) throws SQLException {
+		return Comment.builder()
+			.id(rs.getLong(i + 1))
+			.memberInfoId(rs.getLong(i + 2))
+			.postId(rs.getLong(i + 3))
+			.content(rs.getString(i + 4))
+			.reportCount(rs.getInt(i + 5))
+			.alarm(rs.getBoolean(i + 6))
+			.status(Status.valueOf(rs.getString(i + 7)))
+			.createdAt(rs.getTimestamp(i + 8).toLocalDateTime())
+			.updatedAt(rs.getTimestamp(i + 9).toLocalDateTime())
 			.build();
 	}
