+++ src/main/java/com/blog/tech/domain/member/repository/dao/MemberDao.java
@@ -11,7 +11,6 @@ import java.util.Optional;
 
 import com.blog.tech.domain.member.entity.Member;
 import com.blog.tech.domain.member.repository.ifs.MemberRepository;
-import com.blog.tech.global.utility.db.mapper.MemberMapper;
 
 public class MemberDao implements MemberRepository {
 
@@ -69,7 +68,7 @@ public class MemberDao implements MemberRepository {
 			return Optional.empty();
 		}
 
-		final Member member = MemberMapper.from(rs, 0);
+		final Member member = Member.from(rs, 0);
 		rs.close();
 		pstmt.close();
 
@@ -109,7 +108,7 @@ public class MemberDao implements MemberRepository {
 			return Optional.empty();
 		}
 
-		final Member member = MemberMapper.from(rs, 0);
+		final Member member = Member.from(rs, 0);
 		rs.close();
 		pstmt.close();
