+++ src/main/java/com/blog/tech/domain/post/entity/Category.java
@@ -1,5 +1,8 @@
 package com.blog.tech.domain.post.entity;
 
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -13,5 +16,12 @@ public class Category {
 
 	private Long id;
 	private String name;
+
+	public static Category from(final ResultSet rs, final int i) throws SQLException {
+		return Category.builder()
+			.id(rs.getLong("id"))
+			.name(rs.getString("name"))
+			.build();
+	}
 
 }
