+++ src/main/java/com/blog/tech/domain/post/entity/ConnectHashtag.java
@@ -1,5 +1,8 @@
 package com.blog.tech.domain.post.entity;
 
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -26,4 +29,13 @@ public class ConnectHashtag {
 			.postId(postId)
 			.build();
 	}
+
+	public static ConnectHashtag from(final ResultSet rs, final int i) throws SQLException {
+		return ConnectHashtag.builder()
+			.id(rs.getLong(i + 1))
+			.hashtagId(rs.getLong(i + 2))
+			.postId(rs.getLong(i + 3))
+			.build();
+	}
+
 }
