--- src/main/java/com/blog/tech/global/utility/db/mapper/CategoryMapper.java
@@ -1,17 +1,0 @@
-package com.blog.tech.global.utility.db.mapper;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import com.blog.tech.domain.post.entity.Category;
-
-public class CategoryMapper {
-
-	public static Category from(final ResultSet rs, final int i) throws SQLException {
-		return Category.builder()
-			.id(rs.getLong("id"))
-			.name(rs.getString("name"))
-			.build();
-	}
-
-}
